+++ lib/spellr/interactive_replacement.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'readline'
+require 'reline'
 require_relative 'string_format'
 
 module Spellr
@@ -15,8 +15,8 @@ module Spellr
       @token = token
       @token_highlight = red(token)
       @reporter = reporter
-      Readline.input = reporter.output.stdin
-      Readline.output = reporter.output.stdout
+      Reline.input = reporter.output.stdin
+      Reline.output = reporter.output.stdout
     end
 
     def global_replace
@@ -44,9 +44,9 @@ module Spellr
     end
 
     def prompt_replacement
-      Readline.pre_input_hook = -> { pre_input_hook(token) }
-      prompt = "  Replace #{'all ' if global?}#{token_highlight} with: \e[32m"
-      Readline.readline(prompt)
+      Reline.pre_input_hook = -> { pre_input_hook(token) }
+      prompt = "  Replace #{'all ' if global?}#{token_highlight} with: "
+      Reline.readline(prompt)
     rescue Interrupt
       handle_ctrl_c
     end
@@ -70,19 +70,15 @@ module Spellr
 
     def handle_ctrl_c
       print "\e[0m"
-      reporter.clear_line(4)
+      reporter.clear_line(5)
       reporter.call(token, only_prompt: true)
     end
 
     private
 
     def pre_input_hook(value)
-      Readline.refresh_line
-      Readline.insert_text value.to_s
-      Readline.redisplay
-
-      # Remove the hook right away.
-      Readline.pre_input_hook = nil
+      Reline.insert_text value.to_s
+      Reline.pre_input_hook = nil
     end
 
     def puts(str)