+++ lib/spellr/null_suggester.rb
@@ -1,25 +1,19 @@
 # frozen_string_literal: true
 
 module Spellr
-  class NullSuggester
+  module NullSuggester
     class << self
-      def suggestions(_token)
+      def suggestions(_token, _limit = 0)
         []
       end
 
-      def fast_suggestions(_token)
+      def fast_suggestions(_token, _limit = 0)
         []
       end
 
       def slow?
         true
       end
-    end
-
-    def initialize(_wordlist); end
-
-    def suggestions(_term)
-      []
     end
   end
 end