+++ lib/spellr/output_stubbed.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'stringio'
 require_relative 'output'
 
 module Spellr