+++ lib/spellr/tokenizer.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'stringio'
 require_relative 'token'
 require_relative 'column_location'
 require_relative 'line_location'